Split potentially-patented codecs to a .Codecs extension
It's helpful when pre-installing apps like Audacity that only "safe" codecs
which are royalty-free or patents have expired are included in the base app,
and anything... more adventurous is split out. By default for downloads (and
upgrades) the .Codecs extension is fetched so this has no normal end-user
effect. Build ffmpeg twice, once in a more conservative configuration and once
in an extension provided from /app/lib/codecs. We have to remove the rpath from
the binary otherwise the extension's library path has no effect and the boring
ffmpeg is loaded unconditionally.